Move statsd init to tools.db, rename to resources
