+++ core/schains/monitor/skaled_monitor.py
@@ -29,7 +29,7 @@ from core.schains.config.main import get_number_of_secret_shares
 from core.schains.skaled_status import SkaledStatus
 from core.schains.ssl import ssl_reload_needed
 from tools.configs import SYNC_NODE
-from tools.helper import get_statsd_client
+from tools.resources import get_statsd_client
 from web.models.schain import SChainRecord
