+++ web/models/base.py
@@ -18,7 +18,7 @@
 #   along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
 from peewee import Model
-from tools.db import get_database
+from tools.resources import get_database
 
 
 class BaseModel(Model):
